restructure code logic (doc processing at cli call)
