+++ src/cat_base/cli.py
@@ -4,17 +4,20 @@ import typer
 
 from cat_base.data import (
     create_database,
-    create_database_from_arxiv,
     list_databases,
+    parse_arxiv,
+    parse_documents,
 )
 
 app = typer.Typer()
 
 
 @app.command()
-def say(message: str = "") -> None:
-    """Say a message."""
-    typer.echo(message)
+def list() -> None:
+    """List all databases."""
+    typer.echo("Listing databases:")
+    list_databases()
+    typer.echo("That's it! :)")
 
 
 @app.command()
@@ -24,16 +27,10 @@ def create(
 ) -> None:
     """Create a new database."""
     typer.echo(f"Creating database {database_name} in {pdf_directory}")
-    create_database(database_name, pdf_directory)
+    documents = parse_documents(pdf_directory)
+    create_database(database_name, documents)
+
     typer.echo("Done! :)")
-
-
-@app.command()
-def list() -> None:
-    """List all databases."""
-    typer.echo("Listing databases:")
-
-    list_databases()
 
 
 @app.command()
@@ -46,5 +43,7 @@ def arxiv(
     typer.echo(
         f"Creating database {database_name} from arXiv with keywords {keyword_list}"
     )
-    create_database_from_arxiv(database_name, keyword_list, max_docs)
+    documents = parse_arxiv(keyword_list, max_docs)
+    create_database(database_name, documents)
+
     typer.echo("Done! :)")
