+++ src/cat_base/data/__init__.py
@@ -1,5 +1,8 @@
-from .document_preprocessing import (
+from .database_creation import (
     create_database,
-    create_database_from_arxiv,
     list_databases,
+)
+from .parsing import (
+    parse_arxiv,
+    parse_documents,
 )
